--- src/Components/Main/RecommendWrap/RecommendWrapConst/GnbMenu.js
@@ -1,6 +1,0 @@
-export const TASTE = [
-  { id: 1, content: '매콤한맛' },
-  { id: 2, content: '짭짤한맛' },
-  { id: 3, content: '새콤한맛' },
-  { id: 4, content: '달달한맛' },
-];
